requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-search-documents==11.4.0
backcall==0.2.0
backoff==2.2.1
bidict==0.22.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chainlit==0.7.400
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
fonttools==4.43.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
icecream==2.1.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
langchain==0.0.337
langsmith==0.0.65
Lazify==0.4.0
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
microsoft-kiota-abstractions==1.0.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.0.0
microsoft-kiota-serialization-json==1.0.0
microsoft-kiota-serialization-text==1.0.0
msal==1.24.1
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2
nodeenv==1.8.0
numpy==1.26.1
openai==1.2.3
opencv-python==4.8.1.78
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
portalocker==2.8.2
prisma==0.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyodbc==5.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-image==0.22.0
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
st-chat-message==0.3.8
stack-data==0.6.3
starlette==0.27.0
std-uritemplate==0.0.46
streamlit==1.24.0
streamlit-chat==0.1.1
syncer==2.0.3
tenacity==8.2.3
tifffile==2023.9.26
tiktoken==0.5.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
uptrace==1.20.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==12.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
